requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
Click==7.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.1.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
opencv-python==4.1.2.30
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
pytesseract==0.3.1
python-dateutil==2.8.1
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.13
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
virtualenv==16.7.9
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==2.0.1
